Add Dokka for documentation generation
The build.gradle.kts has been updated to implement and configure Dokka, a library for generating Kotlin documentation. The use of plugins, dependencies, etc. has been added in to aid in the documentation generation process. The dokka version has been included in the libs.versions.toml file. This addition was made to improve code readability and ensure our codebase is well-documented for future reference.