Add Nexus publishing and Git versioning setup to the build script
Updated build.gradle.kts and libs.versions.toml to incorporate git-versioning and nexus-publish plugins for better version management and ease of publishing to Maven Central. Adjusted the project versioning to follow Git tags and added the nexus publishing setup for easier deployment of project artifacts. Additionally, refined the setup for the javadocJar task and signing configuration.