Add GitHub action for publishing snapshots
Created a new GitHub action for publishing snapshot versions of the application. This action is triggered on pushes to the main branch. It begins by checking out the code, then running tests with Gradle. If the tests pass, it utilizes Gradle again to publish the snapshot to the Sonatype Nexus repository. This flow was implemented to automate the delivery of snapshot versions for latest changes in the main branch.