Remove redundant badges and update README structure
Removed redundant maven and sonatype badges and streamlined the README.md structure for better readability. Also, relabeled the badges for clarity.

The changes made to the README file are necessary to improve its readability and help users quickly comprehend the features of the web push library. Sections under 'Usage' are broken down and expanded into smaller sub-sections. This will make it easier to digest information. Text blocks are cleaner and neater with the removal of excess whitespace. The modifications are also aimed to make the documentation more user-friendly and comprehensive, providing detailed instructions on sending notifications and handling VAPID keys.