Update badges positioning and rewrite sentences in README.md
The position of the badges for Maven Central and Sonatype Snapshots have been moved to their respective sections for a more logical and organized look. This ensures they are in a relevant context improving understandability for readers. The term 'WebPushService' has been simplified to 'service' in the instructions to avoid redundancy and improve simplicity.