Update README with language specification and visual aid
This commit updated README.md to specify that the WebPush library is written in Kotlin for clarity and precision. We also added a notification image to visually enhance the explanation for users setting up their service worker. Lastly, we updated the credit for the README illustration to reflect the actual source.