Refactor publishing process in build.gradle.kts
The changes in this commit refactor the way we handle the publishing process in build.gradle.kts.
The changes do not alter functionality but make the code cleaner and more maintainable. Notably, the `nexusPublishing` and `publishing` tasks have been ordered logically, and the `license` section has been moved for better readability.