Add Notification data class
Introduces a Notification data class representing a web push notification to be sent by a WebPushService. Aims to replace the overloaded multi-parameter methods of WebPushService with a single one, deprecating the old ones to avoid repetition of documentation.

Updates the README and BrowserTest to remove references to the deprecated methods.